requirements.txt
fastapi
uvicorn
python-dotenv
paho-mqtt>=1.6.1
ollama
